New notebook of Hyper tuned XGBoost and Mini NN
This notebook was created from the MVP, it expands upon the XGBoost Tree and also implements a new Mini NN with tensorflow. Both have fairly similarly results in terms of MAE, but the XGBoost tree preforms slightly better on the outliers. Both are significantly better with general data, than with outliers.